requirements.txt
python-dotenv
google-cloud-vision
google-auth
google-auth-transport-requests
google-generativeai
Pillow
natsort